+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes qsf-integration project will be documented in this file.
 
 ## [UNRELEASED]
 ### Added
-
+- add userId and sessionId to the request and tracking
 
 ### Changed
 - use the document id as a fallback for tracking returned documents